Style-Finder-Backend/requirements.txt
fastapi
uvicorn
sqlmodel
firebase-admin
google-generativeai
cloudinary
requests
python-multipart
python-dotenv
google-search-results